squash: fixed proptypes & trigger components
