+++ src/Dropdown/__tests__/Dropdown-test.js
@@ -18,26 +18,9 @@ if (React.version.match(/15.[0-5]/)) {
 
 var MockDropdownList = require("./fixtures/MockDropdownList");
 var Dropdown = require("../Dropdown.js");
+var ButtonTrigger = require("../ButtonTrigger.js").default;
 
-class InputTrigger extends React.Component {
-  constructor() {
-    super(...arguments);
-
-    this.handleClick = this.handleClick.bind(this);
-  }
-
-  handleClick(event) {
-    this.props.onAction(event);
-  }
-
-  getSelectedHtml(item) {
-    if (item != null) {
-      return item.selectedHtml || item.html;
-    }
-
-    return "Placeholder";
-  }
-
+class InputTrigger extends ButtonTrigger {
   render() {
     const { className, disabled, selectedItem } = this.props;
 
@@ -47,7 +30,7 @@ class InputTrigger extends React.Component {
         disabled={disabled}
         onClick={this.handleClick}
         readOnly
-        type="input"
+        type="text"
         value={this.getSelectedHtml(selectedItem)}
       />
     );
@@ -284,8 +267,8 @@ describe("Dropdown", function() {
     });
   });
 
-  describe("#getSelectedItem", function() {
-    it("returns item for initialID", function() {
+  describe.only("#getSelectedItem", function() {
+    it.only("returns item for initialID", function() {
       expect(this.instance.getSelectedItem()).toEqual(MockDropdownList[1]);
     });
     it("returns null if nothing is given", function() {
@@ -311,7 +294,7 @@ describe("Dropdown", function() {
           items={MockDropdownList}
           transition={false}
           wrapperClassName="dropdown"
-          trigger={<InputTrigger />}
+          trigger={<InputTrigger placeholder="Placeholder" />}
         />,
         document.body
       );
@@ -332,7 +315,7 @@ describe("Dropdown", function() {
           transition={false}
           initialID="bar"
           wrapperClassName="dropdown"
-          trigger={<InputTrigger />}
+          trigger={<InputTrigger placeholder="Placeholder" />}
         />,
         document.body
       );
